# Backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AutoPulse API",
    "version": "1.0.0",
    "description": "API for managing driving license uploads and operations"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "error": {
            "type": "string",
            "description": "Detailed error information"
          }
        }
      },
...
      "UploadRegistration": {
        "type": "object",
        "properties": {
          "bikeId": {
            "type": "string",
            "example": "685d4f07cc39dcbd6a0171a0"
          },
          "fileId": {
            "type": "string",
            "example": "685dac080cbde4aba4fb7fd6"
          },
          "filename": {
            "type": "string",
            "example": "registration_685d4f07cc39dcbd6a0171a0_1750969352190.pdf"
          },
          "originalName": {
            "type": "string",
            "example": "bike_registration.pdf"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-27T20:22:32.225Z"
          },
          "fileSize": {
            "type": "integer",
            "example": 28310
          },
          "contentType": {
            "type": "string",
            "example": "application/pdf"
          }
        }
      },
      "BikeInfo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "685d4f07cc39dcbd6a0171a0"
          },
          "brand": {
            "type": "string",
            "example": "Suzuki"
          },
          "model": {
            "type": "string",
            "example": "Gixxer"
          },
          "year": {
            "type": "string",
            "example": "2025"
          },
          "registrationNumber": {
            "type": "string",
            "example": "DH-LA-65-9421"
          },
          "odometer": {
            "type": "integer",
            "example": 9500
          }
        }
      },
      "Registration": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "example": "685dac080cbde4aba4fb7fd6"
          },
          "filename": {
            "type": "string",
            "example": "registration_685d4f07cc39dcbd6a0171a0_1750969352190.pdf"
          },
          "originalName": {
            "type": "string",
            "example": "bike_registration.pdf"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-27T20:22:32.225Z"
          },
          "fileSize": {
            "type": "integer",
            "example": 28310
          },
          "contentType": {
            "type": "string",
            "example": "application/pdf"
          }
        }
      },
      "RegistrationStatus": {
        "type": "object",
        "properties": {
          "hasRegistration": {
            "type": "string",
            "description": "Registration file ID if exists, or boolean false if no registration",
            "example": "685dac080cbde4aba4fb7fd6"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-27T20:22:32.225Z"
          },
          "filename": {
            "type": "string",
            "example": "bike_registration.pdf"
          }
        }
      },
      "DeletedRegistration": {
        "type": "object",
        "properties": {
          "fileId": {
            "type": "string",
            "example": "685dac080cbde4aba4fb7fd6"
          },
          "filename": {
            "type": "string",
            "example": "registration_685d4f07cc39dcbd6a0171a0_1750969352190.pdf"
          },
          "originalName": {
            "type": "string",
            "example": "bike_registration.pdf"
          },
          "uploadDate": {
            "type": "string",
            "format": "date-time",
            "example": "2025-06-27T20:22:32.225Z"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/registration/upload/{bikeId}": {
      "post": {
        "summary": "Upload a bike registration document",
        "tags": [
          "Registration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bikeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB ObjectId of the bike",
            "example": "685d4f07cc39dcbd6a0171a0"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "registration": {
                    "type": "string",
                    "format": "binary",
                    "description": "PDF or image file (max 10MB)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration document uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bike registration document uploaded successfully"
                    },
                    "registration": {
                      "$ref": "#/components/schemas/UploadRegistration"
                    }
                  }
                },
                "example": {
                  "message": "Bike registration document uploaded successfully",
                  "registration": {
                    "bikeId": "685d4f07cc39dcbd6a0171a0",
                    "fileId": "685e0e3e16972f595709fe94",
                    "filename": "registration_685d4f07cc39dcbd6a0171a0_1750994494636.pdf",
                    "originalName": "Bike registration.pdf",
                    "uploadDate": "2025-06-27T03:21:34.718Z",
                    "fileSize": 361395,
                    "contentType": "application/pdf"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid file type, size, or missing file)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bike not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/registration/info/{bikeId}": {
      "get": {
        "summary": "Get registration information",
        "tags": [
          "Registration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bikeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB ObjectId of the bike",
            "example": "685d4f07cc39dcbd6a0171a0"
          }
        ],
        "responses": {
          "200": {
            "description": "Registration information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Registration information retrieved successfully"
                    },
                    "bike": {
                      "$ref": "#/components/schemas/BikeInfo"
                    },
                    "registration": {
                      "$ref": "#/components/schemas/Registration"
                    }
                  }
                },
                "example": {
                  "message": "Registration information retrieved successfully",
                  "bike": {
                    "bikeId": "685d4f07cc39dcbd6a0171a0",
                    "brand": "Honda",
                    "model": "CBR600RR",
                    "year": 2023,
                    "plateNumber": "ABC-1234",
                    "owner": {
                      "_id": "685d4cabcc39dcbd6a017198",
                      "name": "John Doe",
                      "email": "john@example.com"
                    }
                  },
                  "registration": {
                    "fileId": "685e0e3e16972f595709fe94",
                    "filename": "registration_685d4f07cc39dcbd6a0171a0_1750994494636.pdf",
                    "originalName": "Bike registration.pdf",
                    "uploadDate": "2025-06-27T03:21:34.718Z",
                    "fileSize": 361395,
                    "contentType": "application/pdf"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid bike ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bike or registration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/registration/download/{bikeId}": {
      "get": {
        "summary": "Download registration file",
        "tags": [
          "Registration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bikeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB ObjectId of the bike",
            "example": "685d4f07cc39dcbd6a0171a0"
          }
        ],
        "responses": {
          "200": {
            "description": "File download",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "attachment; filename=\"filename.ext\""
              },
              "Content-Type": {
                "schema": {
                  "type": "string"
                },
                "description": "MIME type of the file"
              }
            }
          },
          "400": {
            "description": "Invalid bike ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bike or registration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/registration/status/{bikeId}": {
      "get": {
        "summary": "Check registration status",
        "tags": [
          "Registration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bikeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB ObjectId of the bike",
            "example": "685d4f07cc39dcbd6a0171a0"
          }
        ],
        "responses": {
          "200": {
            "description": "Registration status retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Registration status retrieved successfully"
                    },
                    "bike": {
                      "$ref": "#/components/schemas/BikeInfo"
                    },
                    "registrationStatus": {
                      "$ref": "#/components/schemas/RegistrationStatus"
                    }
                  }
                },
                "example": {
                  "message": "Registration status retrieved successfully",
                  "bike": {
                    "bikeId": "685d4f07cc39dcbd6a0171a0",
                    "brand": "Honda",
                    "model": "CBR600RR",
                    "year": 2023,
                    "plateNumber": "ABC-1234",
                    "owner": {
                      "_id": "685d4cabcc39dcbd6a017198",
                      "name": "John Doe",
                      "email": "john@example.com"
                    }
                  },
                  "registrationStatus": {
                    "hasRegistration": true,
                    "uploadDate": "2025-06-27T03:21:34.718Z",
                    "filename": "Bike registration.pdf"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid bike ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bike not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/registration/delete/{bikeId}": {
      "delete": {
        "summary": "Delete registration",
        "tags": [
          "Registration"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "bikeId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "MongoDB ObjectId of the bike",
            "example": "685d4f07cc39dcbd6a0171a0"
          }
        ],
        "responses": {
          "200": {
            "description": "Registration deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bike registration document deleted successfully"
                    },
                    "deletedRegistration": {
                      "$ref": "#/components/schemas/DeletedRegistration"
                    }
                  }
                },
                "example": {
                  "message": "Bike registration document deleted successfully",
                  "deletedRegistration": {
                    "fileId": "685e0e3e16972f595709fe94",
                    "filename": "registration_685d4f07cc39dcbd6a0171a0_1750994494636.pdf",
                    "originalName": "Bike registration.pdf",
                    "uploadDate": "2025-06-27T03:21:34.718Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid bike ID format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Bike or registration not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}